+++ src/main/java/com/khem/appspring/springphoneshop/service/BrandService.java
@@ -8,11 +8,11 @@ import com.khem.appspring.springphoneshop.model.Brand;
 public interface BrandService {
     Brand save(Brand entity);
 
-    Brand getById(Integer id);
+    Brand getById(Long id);
 
-    Brand update(Integer id, Brand  brand);
+    Brand update(Long id, Brand  brand);
 
-    void delete(Integer id);
+    void delete(Long id);
 
     List<Brand> findAll();
 }
